ui/cluster-ui: extract lock-wait combiner into activeStatementUtils
Move the `selectActiveExecutionsCombiner` function from
`activeExecutionsCommon.selectors.ts` into `activeStatementUtils.ts`
as `getActiveExecutionsWithLockWaits`. This makes the combiner logic
available as a plain function (not a selector), which is needed by
the upcoming `useLiveWorkload` SWR hook that combines session and
lock data outside of Redux.

The old selector file is deleted and all references (cluster-ui and
db-console selectors) are updated to use the new function.

Epic: CRDB-58145
Part-of: CRDB-61194
Release note: None
