+++ pkg/ui/workspaces/cluster-ui/src/activeExecutions/index.ts
@@ -5,6 +5,7 @@
 
 export {
   getActiveExecutionsFromSessions,
+  getActiveExecutionsWithLockWaits,
   getContendedExecutionsForTxn,
   getWaitTimeByTxnIDFromLocks,
   getActiveTransaction,
