+++ pkg/ui/workspaces/cluster-ui/src/selectors/index.ts
@@ -46,5 +46,4 @@
 // )
 
 export * from "./common";
-export * from "./activeExecutionsCommon.selectors";
 export * from "./insightsCommon.selectors";
